Pulls wallet extension into separate file.
