+++ tools/walletextension/main/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/obscuronet/obscuro-playground/tools/walletextension"
+)
+
+func main() {
+	cp := walletextension.NewWalletExtension()
+	cp.Serve()
+}
