+++ tools/walletextension/static/index.html
