+++ tools/walletextension/static/javascript.js
